power: oplus_battery_msm8250: Don't rerun DCIN AICL if userspace changes ICL
Currently, DCIN AICL will run automatically if userspace reduces WLS
VOUT. In thermal conditions, when userspace reduces DCIN ICL, we
immediately will run into ILIM, in response to which WLS VOUT will be
reduced. The automatic re-running of DCIN AICL will push ICL back up
again, which is contrary to what userspace wants at that point.

Prevent this from happening. The only reason userspace changes DC ICL is
because of thermal conditions, and we shouldn't negate the actions it
takes as a result of those conditions.

Change-Id: Id0ec44068387bf2ba3c0a0c72f5a8f7a1a527384
Signed-off-by: Guru Das Srinagesh <gurus@codeaurora.org>